frontend,favorite add

+++ app/Http/Controllers/Author/ProductController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Author;
 
 use App\Tag;
 use App\Product;
@@ -22,8 +22,8 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = product::latest()->get();
-        return view('admin.product.index',compact('products'));
+        $products = Auth::user()->products()->get();
+        return view('author.product.index', compact('products'));
     }
 
     /**
@@ -35,7 +35,7 @@ class ProductController extends Controller
     {
         $categories = Category::latest()->get();
         $tags       = Tag::latest()->get();
-        return view('admin.product.create',compact('categories','tags'));
+        return view('author.product.create', compact('categories', 'tags'));
     }
 
     /**
@@ -46,48 +46,48 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'title'=>'required',
+
+        $this->validate($request, [
+            'title' => 'required',
             'image' => 'required|mimes:jpeg,bmp,png,jpg',
-            'categories'=>'required',
-            'tags'=>'required',
-            'body'=>'required',
-         ]);
-         $image = $request->file('image');
-         $slug  = str_slug($request->title);
-         if(isset($image)){
-             $currentDate= Carbon::now()->toDateString();
-             $imagename = $slug.'-'.$currentDate.'-'.uniqid().'.'.$image->getClientOriginalExtension();
-            if(!Storage::disk('public')->exists('product')){
+            'categories' => 'required',
+            'tags' => 'required',
+            'body' => 'required',
+        ]);
+        $image = $request->file('image');
+        $slug  = str_slug($request->title);
+        if (isset($image)) {
+            $currentDate = Carbon::now()->toDateString();
+            $imagename = $slug . '-' . $currentDate . '-' . uniqid() . '.' . $image->getClientOriginalExtension();
+            if (!Storage::disk('public')->exists('product')) {
                 Storage::disk('public')->makeDirectory('product');
-             }
- 
-            $productImage= Image::make($image)->resize(1600,1066)->stream();
-            storage::disk('public')->put('product/'.$imagename,$productImage);
-     
-         }else{
-             $imagename = 'default.png'; // also ok without this else
-         }
-          
-         $product = new Product();
-         $product->user_id = Auth::id();//auth::id()= present authinticated id
-         $product->title   = $request->title;
-         $product->slug    = $slug;
-         $product->image   = $imagename;
-         $product->body   = $request->body;
-         if(isset($request->status)){
-             $product->status = true;
-         }else{
-             $product->status = false;
-         }
- 
-         $product->is_approved = true;//for admin we forcefully make it true , but in author there will something else
+            }
+
+            $productImage = Image::make($image)->resize(1600, 1066)->stream();
+            storage::disk('public')->put('product/' . $imagename, $productImage);
+        } else {
+            $imagename = 'default.png'; // also ok without this else
+        }
+
+        $product = new Product();
+        $product->user_id = Auth::id(); //auth::id()= present authinticated id
+        $product->title   = $request->title;
+        $product->slug    = $slug;
+        $product->image   = $imagename;
+        $product->body   = $request->body;
+        if (isset($request->status)) {
+            $product->status = true;
+        } else {
+            $product->status = false;
+        }
+
+        $product->is_approved = false; //for admin we forcefully make it true , but in author there will something else
         $product->save();
         //$request->tags=tags array by select name
-        $product->categories()->attach($request->categories);//categories()=function in product model
+        $product->categories()->attach($request->categories); //categories()=function in product model
         $product->tags()->attach($request->tags);
-        Toastr::success('product Successfuly saved','success');
-        return redirect()->route('admin.product.index');
+        Toastr::success('product Successfuly saved', 'success');
+        return redirect()->route('author.product.index');
     }
 
     /**
@@ -98,10 +98,13 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
-
-      $categories = Category::latest()->get();
+        if($product->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+        $categories = Category::latest()->get();
         $tags       = Tag::latest()->get();
-        return view('admin.product.show',compact('product','categories','tags'));
+        return view('author.product.show',compact('product','categories','tags'));
     }
 
     /**
@@ -112,9 +115,14 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        if($product->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+        
         $categories = Category::latest()->get();
         $tags       = Tag::latest()->get();
-        return view('admin.product.edit',compact('product','categories','tags'));
+        return view('author.product.edit',compact('product','categories','tags'));
     }
 
     /**
@@ -126,6 +134,11 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+
+        if($product->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         $this->validate($request,[
             'title'=>'required',
             'image' => 'required|mimes:jpeg,bmp,png,jpg',
@@ -166,31 +179,13 @@ class ProductController extends Controller
              $product->status = false;
          }
  
-         $product->is_approved = true;//for admin we forcefully make it true , but in author there will something else
+         $product->is_approved = false;//for admin we forcefully make it true , but in author there will something else
         $product->save();
         //$request->tags=tags array by select name
         $product->categories()->sync($request->categories);//categories()=function in product model
         $product->tags()->sync($request->tags);
         Toastr::success('product Successfuly Updated','success');
-        return redirect()->route('admin.product.index');
-    }
-
-    public function pending(){
-        $products = Product::where('is_approved',false)->get();
-        return view('admin.product.pending',compact('products'));
-    }
-
-    public function approval($id){
-        
-        $product = Product::find($id);
-        if($product->is_approved == false){
-            $product->is_approved = true;
-            $product->save();
-            Toastr::success('Post Successfully Approved','success');
-        }else{
-            Toastr::info('This Post is already approved.','info');
-        }
-        return redirect()->back();
+        return redirect()->route('author.product.index');
     }
 
     /**
@@ -201,6 +196,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        if($product->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         if(storage::disk('public')->exists('product/'.$product->image)){
             Storage::disk('public')->delete('product/'.$product->image);
         }
